MSFT Sync: I18n of errors part 1
Errors happen in a job, so to localize the errors to whatever the user
viewing the page sees, we need to store the un-i18n-ized version in the
database. However, if we do that, the messages won't be extracted to the
i18n yml file; so we still need to call I18n.t (we do this in #public_error)
with a constant string so the extractor can pick them out.
I'm not sure if jobs are even run in a non-en locale, but in case they
are, we temporarily set the locale to 'en' before running that I18n.t()
for that error to make sure we store the un-i18n-ized version.

TODO follow up commits:
* change more errors to PublicErrors
* GracefulCancelError can be a subclass of PublicError

Test plan:
- have a course with no owners (teachers/tas/designers) and try to sync.
- check the last_error field on the MicrosoftSync::Group model. It
  should be a JSON hash with appropriate public_error_message and no
  public_interpolated_values.
- check the UI (course integrations settings) to make sure the error
  shows up there (just the public message, not JSON)
- Add the following to es.config/locales/yml:
  microsoft_365_allows_a_maximum_of_max_owners_in_a__b448f5de: "Microsoft 365 permite como máximo %{max} administradores de un Microsoft Team."
- change MembershipDiff::MAX_ENROLLMENT_OWNERS to 2. restart Canvas.
  Add two owners in a course and try to sync.
- check the last_error. It should have a public_message and separate
  public_interpolated_values with "max": 2.
- check the UI to make sure the error message shows up with
  the interpolations.
- change your language to Español
  (/profile/settings -> Edit Settings) and reload the course settings
  integration page. The error message should now show up in Spanish.
  ¡Empieza la fiesta!

refs INTEROP-6739
flag=microsoft_group_enrollments_syncing

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/269064
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>

+++ lib/microsoft_sync/syncer_steps.rb
@@ -66,12 +66,47 @@ module MicrosoftSync
       include Errors::GracefulCancelErrorMixin
     end
 
-    class MissingOwners < SyncCanceled; end
+    class MissingOwners < SyncCanceled
+      def self.public_message
+        I18n.t 'A Microsoft 365 Group must have owners, and no users ' \
+          'corresponding to the instructors of the Canvas course could be found on the ' \
+          'Microsoft side.'
+      end
+    end
+
     # Can happen when User disables sync on account-level when jobs are running:
-    class TenantMissingOrSyncDisabled < SyncCanceled; end
-    # Can happen when the Course has more then 25k members's enrolled or 100
-    # owner's enrolled
-    class MaxEnrollmentsReached < SyncCanceled; end
+    class TenantMissingOrSyncDisabled < SyncCanceled
+      def self.public_message
+        I18n.t 'Tenant missing or sync disabled. ' \
+          'Check the Microsoft sync integration settings for the course and account.'
+      end
+    end
+
+    class MultipleEducationClasses < SyncCanceled
+      def self.public_message
+        I18n.t 'Multiple Microsoft education classes already exist for the course.'
+      end
+    end
+
+    class MaxMemberEnrollmentsReached < SyncCanceled
+      def self.public_message
+        I18n.t 'Microsoft 365 allows a maximum of %{max} members in a team.'
+      end
+
+      def public_interpolated_values
+        {max: MAX_ENROLLMENT_MEMBERS}
+      end
+    end
+
+    class MaxOwnerEnrollmentsReached < SyncCanceled
+      def self.public_message
+        I18n.t 'Microsoft 365 allows a maximum of %{max} owners in a team.'
+      end
+
+      def public_interpolated_values
+        {max: MAX_ENROLLMENT_OWNERS}
+      end
+    end
 
     attr_reader :group
     delegate :course, to: :group
@@ -116,8 +151,8 @@ module MicrosoftSync
     # don't want to delete PartialSyncChanges corresponding to enrollments not
     # yet replicated.)
     def step_full_sync_prerequisites(_mem_data, _job_state_data)
-      raise_max_enrollment_members_reached if max_enrollment_members_reached?
-      raise_max_enrollment_owners_reached if max_enrollment_owners_reached?
+      raise MaxMemberEnrollmentsReached if max_enrollment_members_reached?
+      raise MaxOwnerEnrollmentsReached if max_enrollment_owners_reached?
 
       PartialSyncChange.delete_all_replicated_to_secondary_for_course(course.id)
 
@@ -133,10 +168,7 @@ module MicrosoftSync
         return StateMachineJob::NextStep.new(:step_ensure_enrollments_user_mappings_filled)
       end
 
-      if remote_ids.length > 1
-        raise MicrosoftSync::Errors::InvalidRemoteState, \
-              "Multiple Microsoft education classes exist for the course."
-      end
+      raise MultipleEducationClasses if remote_ids.length > 1
 
       # Create a group if needed. If there is already a group but we do not
       # have it in the Group record, use it but first update it with course
@@ -231,14 +263,10 @@ module MicrosoftSync
     def step_execute_diff(diff, _job_state_data)
       # TODO: If there are no instructor enrollments, we actually want to
       # remove the group on the Microsoft side (INTEROP-6672)
-      if diff.local_owners.empty?
-        raise MissingOwners, 'A Microsoft 365 Group must have owners, and no users ' \
-          'corresponding to the instructors of the Canvas course could be found on the ' \
-          'Microsoft side.'
-      end
+      raise MissingOwners if diff.local_owners.empty?
 
-      raise_max_enrollment_members_reached if diff.max_enrollment_members_reached?
-      raise_max_enrollment_owners_reached if diff.max_enrollment_owners_reached?
+      raise MaxMemberEnrollmentsReached if diff.max_enrollment_members_reached?
+      raise MaxOwnerEnrollmentsReached if diff.max_enrollment_owners_reached?
 
       execute_diff(diff)
 
@@ -419,16 +447,6 @@ module MicrosoftSync
         .limit(MAX_ENROLLMENT_OWNERS + 1)
         .distinct
         .count > MAX_ENROLLMENT_OWNERS
-    end
-
-    def raise_max_enrollment_members_reached
-      raise MaxEnrollmentsReached, "Microsoft 365 allows a maximum of " \
-          "#{MAX_ENROLLMENT_MEMBERS} members in a team."
-    end
-
-    def raise_max_enrollment_owners_reached
-      raise MaxEnrollmentsReached, "Microsoft 365 allows a maximum of " \
-          "#{MAX_ENROLLMENT_OWNERS} owners in a team."
     end
   end
 end
